removed unnecessary files and consolidated similar logic
