+++ app/logic/getServiceLearningCoursesData.py
@@ -2,13 +2,13 @@ from app.models.course import Course
 from app.models.user import User
 from app.models.courseInstructor import CourseInstructor
 
-def getProposalData(user):
-    """Returns dictionary with data used to populate SL proposal table"""
+def getServiceLearningCoursesData(user):
+    """Returns dictionary with data used to populate service learning proposal table"""
     courses = (Course.select()
                      .where(CourseInstructor.user==user)
                      .join(CourseInstructor)
                      .order_by(Course.id))
-    courseDict = {} #a dictionary of dictionaries for better readability
+    courseDict = {}
     for course in courses:
         otherInstructors = (CourseInstructor.select().where(CourseInstructor.course==course))
         faculty = [f"{instructor.user.firstName} {instructor.user.lastName}" for instructor in otherInstructors]
