+++ app/static/js/slcProposal.js
@@ -19,14 +19,6 @@ $("#instructorTable").on("click", "#remove", function() {
    $(this).closest("tr").remove();
 });
 
-$('td').click( function() {
-     var $this = $(this);
-     if ($this.text() == 'X')
-     {
-         $this.siblings().find('input').val('');
-     }
-   });
-
 let courseInstructors = []
 $('#continue').on("click", function() {
   $("#instructorTable tr").each(function(a, b) {
@@ -44,6 +36,13 @@ $('#continue').on("click", function() {
   });
 });
 
+$('td').click( function() {
+     var $this = $(this);
+     if ($this.text() == 'X')
+     {
+         $this.siblings().find('input').val('');
+     }
+   });
 
 // -----------------TODO: Rewrite this functionality to make it more general.
 $(document).ready ( function() {
