--- app/templates/admin/createSLProposalTable.html
@@ -1,66 +1,0 @@
-{% extends "base.html"%}
-
-{% block scripts %}
-  {{super()}}
-  <script src="/static/js/changeProposalAction.js"></script>
-{% endblock %}
-{% block styles %}
-  {{super()}}
-{% endblock %}
-{% block app_content %}
-<h1 class="text-center">{{g.current_user.firstName}} {{g.current_user.lastName}}</h1>
-<div>
-  <table class="table table-striped">
-    <thead>
-      <tr>
-        <th scope="col">Courses</th>
-        <th scope="col">Faculty</th>
-        <th scope="col">Term</th>
-        <th scope="col">Status</th>
-        <th scope="col">Action</th>
-      </tr>
-    </thead>
-    <tbody>
-        {% for course in courseDict %}
-          <tr>
-            <td>{{courseDict[course]['name']}}</td>
-            <td>{{courseDict[course]['faculty']|join(", ")}}</td>
-            <td>{{courseDict[course]['term'].description}}</td>
-            <td>{{courseDict[course]['status']}}</td>
-            <td>
-
-              <select class="form-select" id="{{courseDict[course]['id']}}" onchange='changeAction(this)'>
-                <option value="---" disabled selected>Select Action</option>
-                <option value="Renew">Renew</option>
-                <option value="View">View</option>
-                <option value="Withdraw">Withdraw</option>
-                <option value="Edit">Edit</option>
-              </select>
-            </td>
-          </tr>
-        {% endfor %}
-    </tbody>
-  </table>
-<!-- Modal -->
-<div class="modal fade" id="withdrawModal" tabindex="-1" aria-labelledby="withdrawModalLabel" aria-hidden="true">
-  <div class="modal-dialog">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title" id="withdrawModalLabel">Remove Course</h5>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-      </div>
-      <div class="modal-body">
-        Are you sure you want to withdraw this SL Course? <br>
-        You will have to create or renew this proposal again.
-        <input id="courseToRemove" type="hidden" value="">
-      </div>
-      <div class="modal-footer">
-        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
-        <button type="button" class="btn btn-primary" onclick="withdraw()">Withdraw</button>
-      </div>
-    </div>
-  </div>
-</div>
-</div>
-
-{% endblock %}
