--- app/templates/main/slcProposal.html
@@ -1,95 +1,0 @@
-{% extends "base.html" %}
-{% block scripts %}
-
-  {{super()}}
-  <script type="text/javascript"></script>
-  <style media="screen"></style>
-
-{% endblock %}
-
-{% block app_content %}
-<h1> sdkfjalskdjfkjf</h1>
-  <h2 class="title" align="center"> Service-Learning Course Designation Active Learning Experience (ALE) Requirements met through Service-Learning </h2>
-  <h3 class="subtitle"> CRITERIA AND PROPOSAL FORM</h3>
-  <p>
-    SERVICE-LEARNING at Berea College is aneducational experience based upon a collaborative partnership between Collegeand community. Learning through service enables students to apply academic knowledge and critical-thinking skills to meet genuine community needs. Through reflection and assessment, students gain deeper understanding of course content and the importance of civic engagement.
-    <br>
-    All Designated Service-Learning Courses will meet the Active Learning Experience (ALE) Requirement.
-    <br>
-    If you have questions about the criteria for Service-Learning Courses, please call Ashley Cochrane at x. 3605.
-  </p>
-
-  <p class="section-title">Criteria for the Service-Learning Course Designation/ALE Requirement met through Service-Learning</p>
-
-  <h3 class="section-title">Student Learning Outcomes </h3>
-  <div>
-    <p>Students will be able to:</p>
-    <ul>
-      <li>
-        Apply academic knowledge and critical-thinking skills to meet genuine community needs.
-      </li>
-      <li>
-        Demonstrate deeper understanding of specific course learning objectives appropriate to the experience.
-      </li>
-      <li>
-        Reflect critically on multiple facets of their experience, on the connections between theory and practice, andbetween learning inside and outside the classroom
-      </li>
-      <li>
-        Faculty will identify one or more course learning objectives that students will address through their service work in the community.
-      </li>
-      <li>
-        Together, the community partner and faculty member will identify community needs and will structure a service-learning experience designed to address those needs
-      </li>
-    </ul>
-  </div>
-
-  <h3> Service-Learning Experience </h3>
-
-  <div>
-    <ul>
-      <li>
-        Service-learning activities will be integrated into course work and will serve as the pedagogical method for the specified service-learning course objective(s). As such, all students in the course will participate in the service-learning activities.
-      </li>
-      <li>
-        The service-learning experience will be substantial and sustained throughout the course, involving active student engagement.
-      </li>
-    </ul>
-  </div>
-
-  <h3> Community Partnerships </h3>
-  <div>
-    <ul>
-      <li>
-        Faculty will identify and designate an appropriate community partner(s) in order for course learning objectives to be met.
-      </li>
-      <li>
-        All partners will be involved in the planning, implementation, and evaluation of the structured service-learning experience.
-      </li>
-    </ul>
-  </div>
-
-  <h3> Reflection and Assessment </h3>
-  <div>
-    <ul>
-      <li>
-        Faculty will infuse critical reflection throughout the service-learning experience; the reflection will create the connection among service work, course objectives, and student learning outcomes.
-      </li>
-      <li>
-        Evaluative criteria for service-learning reflection will be included in the syllabi, including the percentage that reflection will represent in the final grade.
-      </li>
-    </ul>
-  </div>
-
-  <h3> Summary Presentation and Assessment </h3>
-  <div>
-    <ul>
-      <li>
-        The service-learning experience will result in a summary presentation, performance, exhibition, and/or othersynthesizing project completed by the students engaged in the experience.
-      </li>
-      <li>
-        The service-learning experience will result in a summary presentation, performance, exhibition, and/or othersynthesizing project completed by the students engaged in the experience.
-      </li>
-    </ul>
-  </div>
-
- {% endblock %}
