--- app/templates/serviceLearning/serviceCourseManagment.html
@@ -1,22 +1,0 @@
-{% extends "base.html"%}
-{% block scripts %}
-  {{super()}}
-  <script src="/static/js/createEvents.js"></script>
-  <script src="//cdnjs.cloudflare.com/ajax/libs/timepicker/1.3.5/jquery.timepicker.min.js"></script>
-
-{% endblock %}
-
-{% block styles %}
-  {{super()}}
-    <link rel="stylesheet" href="//cdnjs.cloudflare.com/ajax/libs/timepicker/1.3.5/jquery.timepicker.min.css">
-
-{% endblock %}
-
-{% block app_content %}
-<div class="ms-5">
-  <a class="btn btn-primary" href="/slcGuidelines" role="button">Create Course Proposal + </a>
-</div>
-<div> Welcome to Service Course Managment! This site is currently under Construction </div>
-
- </div>>
-{% endblock %}
