+++ app/templates/serviceLearning/slcGuidelines.html
@@ -85,18 +85,22 @@
   <div>
     <ul>
       <li>
-        The service-learning experience will result in a summary presentation, performance, exhibition, and/or othersynthesizing project completed by the students engaged in the experience.
-      </li>
-      <li>
-        The service-learning experience will result in a summary presentation, performance, exhibition, and/or othersynthesizing project completed by the students engaged in the experience.
+        The service-learning experience will result in a summary presentation, performance, exhibition, and/or other synthesizing project completed by the students engaged in the experience.
       </li>
     </ul>
   </div>
-
-  <div  align="right">
-    <label for="nextButton">(1/3)</label>
-    <a role="button" class="btn btn-primary" type="button" id="nextButton" href="/slcProposal"> Next </a>
+  <div class="row align-items-end pt-5">
+    <div class="col">
+      <a
+        href="/serviceCourseManagement"
+        type="button"
+        class="btn btn-danger col">
+        Cancel
+      </a>
+    </div>
+    <div class="col" align="right">
+      <label for="nextButton">(1/3)</label>
+      <a role="button" class="btn btn-primary" type="button" id="nextButton" href="/slcProposal">Next</a>
+    </div>
   </div>
-
-
  {% endblock %}
